Add node to substates conversion
