+++ radix-engine/src/engine/kernel.rs
@@ -1288,8 +1288,8 @@ where
         assert!(taken_nodes.len() == 1);
         let root_node = taken_nodes.into_values().nth(0).unwrap();
 
-        for (id, node) in root_node.to_nodes(node_id) {
-            self.track.put_node(id, node);
+        for (id, substate) in node_to_substates(root_node.to_nodes(node_id)) {
+            self.track.put_substate(id, substate);
         }
 
         for m in &mut self.modules {
