+++ radix-engine/src/model/substates/substate.rs
@@ -1,3 +1,4 @@
+use crate::engine::HeapRENode;
 use crate::model::*;
 use crate::types::*;
 
@@ -236,4 +237,76 @@ impl Into<VaultSubstate> for Substate {
             panic!("Not a vault");
         }
     }
+}
+
+pub fn node_to_substates(nodes: HashMap<RENodeId, HeapRENode>) -> HashMap<SubstateId, Substate> {
+    let mut substates = HashMap::<SubstateId, Substate>::new();
+    for (id, node) in nodes {
+        match node {
+            HeapRENode::Bucket(_) => panic!("Unexpected"),
+            HeapRENode::Proof(_) => panic!("Unexpected"),
+            HeapRENode::Vault(vault) => {
+                let resource = vault
+                    .resource()
+                    .expect("Vault should be liquid at end of successful transaction");
+                let substate = VaultSubstate(resource);
+                let substate_id = match id {
+                    RENodeId::Vault(vault_id) => SubstateId::Vault(vault_id),
+                    _ => panic!("Unexpected"),
+                };
+                substates.insert(substate_id, substate.into());
+            }
+            HeapRENode::KeyValueStore(_) => panic!("Unexpected"),
+            HeapRENode::Component(component) => {
+                let address = match id {
+                    RENodeId::Component(address) => address,
+                    _ => panic!("Unexpected"),
+                };
+                substates.insert(SubstateId::ComponentInfo(address), component.info.into());
+                substates.insert(SubstateId::ComponentState(address), component.state.into());
+            }
+            HeapRENode::Worktop(_) => panic!("Unexpected"),
+            HeapRENode::Package(package) => {
+                let address = match id {
+                    RENodeId::Package(address) => address,
+                    _ => panic!("Unexpected"),
+                };
+                let substate = PackageSubstate {
+                    code: package.code,
+                    blueprint_abis: package.blueprint_abis,
+                };
+                substates.insert(SubstateId::Package(address), substate.into());
+            }
+            HeapRENode::Resource(resource_manager, _) => {
+                let address = match id {
+                    RENodeId::ResourceManager(address) => address,
+                    _ => panic!("Unexpected"),
+                };
+                let substate = ResourceManagerSubstate {
+                    resource_type: resource_manager.resource_type,
+                    metadata: resource_manager.metadata,
+                    method_table: resource_manager.method_table,
+                    vault_method_table: resource_manager.vault_method_table,
+                    bucket_method_table: resource_manager.bucket_method_table,
+                    authorization: resource_manager.authorization,
+                    total_supply: resource_manager.total_supply,
+                };
+                substates.insert(SubstateId::ResourceManager(address), substate.into());
+            }
+            HeapRENode::System(system) => {
+                let address = match id {
+                    RENodeId::System(address) => address,
+                    _ => panic!("Unexpected"),
+                };
+                substates.insert(
+                    SubstateId::System(address),
+                    SystemSubstate {
+                        epoch: system.epoch,
+                    }
+                    .into(),
+                );
+            }
+        }
+    }
+    substates
 }
